split views file
