+++ apps/experiments/views/experiment.py
@@ -1,7 +1,5 @@
-import json
 import logging
-from collections import defaultdict
-from datetime import datetime, timedelta
+from datetime import datetime
 from typing import Optional
 from urllib.parse import quote
 
@@ -13,7 +11,7 @@ from django.contrib import messages
 from django.contrib.auth.decorators import user_passes_test
 from django.core.exceptions import ValidationError
 from django.db.models import Q
-from django.http import Http404, HttpResponse, HttpResponseRedirect, JsonResponse
+from django.http import Http404, HttpResponse, HttpResponseRedirect
 from django.shortcuts import get_object_or_404
 from django.template.loader import render_to_string
 from django.template.response import TemplateResponse
@@ -21,8 +19,6 @@ from django.urls import reverse
 from django.utils import timezone
 from django.utils.safestring import mark_safe
 from django.views.decorators.http import require_POST
-from django.views.generic import CreateView, UpdateView
-from django_tables2 import SingleTableView
 
 from apps.channels.models import ChannelSession, ExperimentChannel
 from apps.chat.models import ChatMessage
@@ -31,20 +27,10 @@ from apps.experiments.email import send_experiment_invitation
 from apps.experiments.export import experiment_to_csv
 from apps.experiments.forms import ConsentForm, ExperimentInvitationForm, SurveyForm
 from apps.experiments.helpers import get_real_user_or_none
-from apps.experiments.models import (
-    Experiment,
-    ExperimentSession,
-    Participant,
-    Prompt,
-    PromptBuilderHistory,
-    SafetyLayer,
-    SessionStatus,
-    SourceMaterial,
-)
-from apps.experiments.tasks import get_prompt_builder_response_task, get_response_for_webchat_task
+from apps.experiments.models import Experiment, ExperimentSession, Participant, SessionStatus
+from apps.experiments.tasks import get_response_for_webchat_task
 from apps.teams.decorators import login_and_team_required, team_admin_required
 from apps.users.models import CustomUser
-from templates.experiments.tables import SafetyLayerTable
 
 
 @login_and_team_required
@@ -58,201 +44,6 @@ def experiments_home(request, team_slug: str):
             "active_tab": "experiments",
         },
     )
-
-
-@login_and_team_required
-def safety_layer_home(request, team_slug: str):
-    return TemplateResponse(
-        request,
-        "generic/object_home.html",
-        {
-            "active_tab": "safety_layers",
-            "title": "Safety Layers",
-            "new_object_url": reverse("experiments:safety_new", args=[team_slug]),
-            "table_url": reverse("experiments:safety_table", args=[team_slug]),
-        },
-    )
-
-
-class SafetyLayerTableView(SingleTableView):
-    model = SafetyLayer
-    paginate_by = 25
-    table_class = SafetyLayerTable
-    template_name = "table/single_table.html"
-
-
-class CreateSafetyLayer(CreateView):
-    model = SafetyLayer
-    fields = [
-        "prompt",
-        "messages_to_review",
-        "default_response_to_user",
-        "prompt_to_bot",
-    ]
-    template_name = "generic/object_form.html"
-    extra_context = {
-        "title": "Create Safety Layer",
-        "button_text": "Create",
-        "active_tab": "safety_layers",
-    }
-
-    def get_success_url(self):
-        return reverse("experiments:safety_home", args=[self.request.team.slug])
-
-    def form_valid(self, form):
-        form.instance.owner = self.request.user
-        return super().form_valid(form)
-
-
-class EditSafetyLayer(UpdateView):
-    model = SafetyLayer
-    fields = [
-        "prompt",
-        "messages_to_review",
-        "default_response_to_user",
-        "prompt_to_bot",
-    ]
-    template_name = "generic/object_form.html"
-    extra_context = {
-        "title": "Update Safety Layer",
-        "button_text": "Update",
-        "active_tab": "safety_layers",
-    }
-
-    def get_form(self):
-        form = super().get_form()
-        form.fields["prompt"].disabled = True
-        return form
-
-    def get_success_url(self):
-        return reverse("experiments:safety_home", args=[self.request.team.slug])
-
-
-@login_and_team_required
-def delete_safety_layer(request, team_slug: str, pk: int):
-    safety_layer = get_object_or_404(SafetyLayer, id=pk)
-    safety_layer.delete()
-    return HttpResponse()
-
-
-@login_and_team_required
-def prompt_builder_load_prompts(request, team_slug: str):
-    prompts = Prompt.objects.all()
-    prompts_list = list(prompts.values())
-
-    return TemplateResponse(
-        request,
-        "experiments/prompts_list.html",
-        {
-            "prompts": prompts_list,
-        },
-    )
-
-
-@login_and_team_required
-def prompt_builder_load_source_material(request, team_slug: str):
-    source_material = SourceMaterial.objects.all()
-    source_material_list = list(source_material.values())
-
-    return TemplateResponse(
-        request,
-        "experiments/source_material_list.html",
-        {
-            "source_materials": source_material_list,
-        },
-    )
-
-
-@login_and_team_required
-def experiments_prompt_builder(request, team_slug: str):
-    prompts = Prompt.objects.order_by("-created_at").all()
-    prompts_list = list(prompts.values())
-
-    return TemplateResponse(
-        request,
-        "experiments/prompt_builder.html",
-        {
-            "prompts": prompts_list,
-            "active_tab": "prompt_builder",
-        },
-    )
-
-
-@require_POST
-@login_and_team_required
-def experiments_prompt_builder_get_message(request, team_slug: str):
-    data_json = request.body.decode("utf-8")
-    user = get_real_user_or_none(request.user)
-    result = get_prompt_builder_response_task.delay(user.id, data_json)
-    return JsonResponse({"task_id": result.task_id})
-
-
-def get_prompt_builder_message_response(request, team_slug: str):
-    task_id = request.GET.get("task_id")
-    progress = Progress(AsyncResult(task_id))
-    return JsonResponse(
-        {
-            "task_id": task_id,
-            "progress": progress.get_info(),
-        },
-    )
-
-
-@login_and_team_required
-def get_prompt_builder_history(request, team_slug: str):
-    # Fetch history for the request user limited to last 30 days
-    thirty_days_ago = timezone.now() - timedelta(days=30)
-    histories = PromptBuilderHistory.objects.filter(owner=request.user, created_at__gte=thirty_days_ago).order_by(
-        "-created_at"
-    )
-
-    # Initialize temporary output format
-    output_temp = defaultdict(list)
-
-    history_id = 0  # Initial history_id for the oldest item
-    for history in histories:
-        # Adding history_id to JSON history object
-        history_data = history.history
-        history_data["history_id"] = history_id
-
-        # Customizing the output for each history
-        event = {
-            "history_id": history_id,
-            "time": history.created_at.strftime("%H:%M"),
-            "preview": history_data.get("preview", ""),
-            "sourceMaterialName": history_data.get("sourceMaterialName", "None"),
-            "sourceMaterialID": history_data.get("sourceMaterialID", -1),
-            "temperature": history_data.get("temperature", 0.7),
-            "prompt": history_data.get("prompt", ""),
-            "inputFormatter": history_data.get("inputFormatter", ""),
-            "model": history_data.get("model", "gpt-4"),
-            "messages": history_data.get("messages", []),
-        }
-
-        # Populating the temporary output dictionary
-        output_temp[history.created_at.date()].append(event)
-
-        history_id += 1  # Incrementing history_id for each newer history
-
-    # Convert to the final desired output format
-    output_list = [
-        {"date": date_obj.strftime("%A %d %b %Y"), "events": events} for date_obj, events in output_temp.items()
-    ]
-
-    return JsonResponse(output_list, safe=False)
-
-
-@login_and_team_required
-@team_admin_required
-def prompt_builder_start_save_process(request, team_slug: str):
-    # Get your long data
-    long_data = json.loads(request.body)
-
-    # Save it in session
-    request.session["long_data"] = long_data
-
-    # Redirect to admin add page
-    return JsonResponse({"redirect_url": reverse("admin:experiments_prompt_add")})
 
 
 @login_and_team_required
